Texture includes spritesheet information, fixed scale filtering issue, level parser fixed to load level

+++ IgniteEngineProject/IgniteEngine/Core/Input/Keycode.h
@@ -6,65 +6,65 @@ namespace ignite
 namespace Keycode
 {
 
-constexpr uint16_t KEY_A      = 0x00000061u;
-constexpr uint16_t KEY_B      = 0x00000062u;
-constexpr uint16_t KEY_C      = 0x00000063u;
-constexpr uint16_t KEY_D      = 0x00000064u;
-constexpr uint16_t KEY_E      = 0x00000065u;
-constexpr uint16_t KEY_F      = 0x00000066u;
-constexpr uint16_t KEY_G      = 0x00000067u;
-constexpr uint16_t KEY_H      = 0x00000068u;
-constexpr uint16_t KEY_I      = 0x00000069u;
-constexpr uint16_t KEY_J      = 0x0000006au;
-constexpr uint16_t KEY_K      = 0x0000006bu;
-constexpr uint16_t KEY_L      = 0x0000006cu;
-constexpr uint16_t KEY_M      = 0x0000006du;
-constexpr uint16_t KEY_N      = 0x0000006eu;
-constexpr uint16_t KEY_O      = 0x0000006fu;
-constexpr uint16_t KEY_P      = 0x00000070u;
-constexpr uint16_t KEY_Q      = 0x00000071u;
-constexpr uint16_t KEY_R      = 0x00000072u;
-constexpr uint16_t KEY_S      = 0x00000073u;
-constexpr uint16_t KEY_T      = 0x00000074u;
-constexpr uint16_t KEY_U      = 0x00000075u;
-constexpr uint16_t KEY_V      = 0x00000076u;
-constexpr uint16_t KEY_W      = 0x00000077u;
-constexpr uint16_t KEY_X      = 0x00000078u;
-constexpr uint16_t KEY_Y      = 0x00000079u;
-constexpr uint16_t KEY_Z      = 0x0000007au;
-constexpr uint16_t KEY_0      = 0x00000030u;
-constexpr uint16_t KEY_1      = 0x00000031u;
-constexpr uint16_t KEY_2      = 0x00000032u;
-constexpr uint16_t KEY_3      = 0x00000033u;
-constexpr uint16_t KEY_4      = 0x00000034u;
-constexpr uint16_t KEY_5      = 0x00000035u;
-constexpr uint16_t KEY_6      = 0x00000036u;
-constexpr uint16_t KEY_7      = 0x00000037u;
-constexpr uint16_t KEY_8      = 0x00000038u;
-constexpr uint16_t KEY_9      = 0x00000039u;
-constexpr uint16_t KEY_ESCAPE = 0x0000001bu;
-constexpr uint16_t KEY_SPACE  = 0x00000020u;
-constexpr uint16_t KEY_RIGHT  = 0x4000004fu;
-constexpr uint16_t KEY_LEFT   = 0x40000050u;
-constexpr uint16_t KEY_DOWN   = 0x40000051u;
-constexpr uint16_t KEY_UP     = 0x40000052u;
-constexpr uint16_t KEY_LCTRL  = 0x400000e0u;
-constexpr uint16_t KEY_LSHIFT = 0x400000e1u;
-constexpr uint16_t KEY_LALT   = 0x400000e2u;
-constexpr uint16_t KEY_RCTRL  = 0x400000e4u;
-constexpr uint16_t KEY_RSHIFT = 0x400000e5u;
-constexpr uint16_t KEY_RALT   = 0x400000e6u;
+constexpr uint32_t KEY_A      = 0x00000061u;
+constexpr uint32_t KEY_B      = 0x00000062u;
+constexpr uint32_t KEY_C      = 0x00000063u;
+constexpr uint32_t KEY_D      = 0x00000064u;
+constexpr uint32_t KEY_E      = 0x00000065u;
+constexpr uint32_t KEY_F      = 0x00000066u;
+constexpr uint32_t KEY_G      = 0x00000067u;
+constexpr uint32_t KEY_H      = 0x00000068u;
+constexpr uint32_t KEY_I      = 0x00000069u;
+constexpr uint32_t KEY_J      = 0x0000006au;
+constexpr uint32_t KEY_K      = 0x0000006bu;
+constexpr uint32_t KEY_L      = 0x0000006cu;
+constexpr uint32_t KEY_M      = 0x0000006du;
+constexpr uint32_t KEY_N      = 0x0000006eu;
+constexpr uint32_t KEY_O      = 0x0000006fu;
+constexpr uint32_t KEY_P      = 0x00000070u;
+constexpr uint32_t KEY_Q      = 0x00000071u;
+constexpr uint32_t KEY_R      = 0x00000072u;
+constexpr uint32_t KEY_S      = 0x00000073u;
+constexpr uint32_t KEY_T      = 0x00000074u;
+constexpr uint32_t KEY_U      = 0x00000075u;
+constexpr uint32_t KEY_V      = 0x00000076u;
+constexpr uint32_t KEY_W      = 0x00000077u;
+constexpr uint32_t KEY_X      = 0x00000078u;
+constexpr uint32_t KEY_Y      = 0x00000079u;
+constexpr uint32_t KEY_Z      = 0x0000007au;
+constexpr uint32_t KEY_0      = 0x00000030u;
+constexpr uint32_t KEY_1      = 0x00000031u;
+constexpr uint32_t KEY_2      = 0x00000032u;
+constexpr uint32_t KEY_3      = 0x00000033u;
+constexpr uint32_t KEY_4      = 0x00000034u;
+constexpr uint32_t KEY_5      = 0x00000035u;
+constexpr uint32_t KEY_6      = 0x00000036u;
+constexpr uint32_t KEY_7      = 0x00000037u;
+constexpr uint32_t KEY_8      = 0x00000038u;
+constexpr uint32_t KEY_9      = 0x00000039u;
+constexpr uint32_t KEY_ESCAPE = 0x0000001bu;
+constexpr uint32_t KEY_SPACE  = 0x00000020u;
+constexpr uint32_t KEY_RIGHT  = 0x4000004fu;
+constexpr uint32_t KEY_LEFT   = 0x40000050u;
+constexpr uint32_t KEY_DOWN   = 0x40000051u;
+constexpr uint32_t KEY_UP     = 0x40000052u;
+constexpr uint32_t KEY_LCTRL  = 0x400000e0u;
+constexpr uint32_t KEY_LSHIFT = 0x400000e1u;
+constexpr uint32_t KEY_LALT   = 0x400000e2u;
+constexpr uint32_t KEY_RCTRL  = 0x400000e4u;
+constexpr uint32_t KEY_RSHIFT = 0x400000e5u;
+constexpr uint32_t KEY_RALT   = 0x400000e6u;
 
 } // Namespace ignite::keycode.
 
 namespace Mouse
 {
 
-constexpr uint16_t BUTTON_LEFT   = 1;
-constexpr uint16_t BUTTON_MIDDLE = 2;
-constexpr uint16_t BUTTON_RIGHT  = 3;
-constexpr uint16_t BUTTON_X1     = 4;
-constexpr uint16_t BUTTON_X2     = 5;
+constexpr uint32_t BUTTON_LEFT   = 1;
+constexpr uint32_t BUTTON_MIDDLE = 2;
+constexpr uint32_t BUTTON_RIGHT  = 3;
+constexpr uint32_t BUTTON_X1     = 4;
+constexpr uint32_t BUTTON_X2     = 5;
 
 } // Namespace ignite::mouse.
